update code --- 用户行为-关注/取消关注 -存储在redis中

+++ heima-leadnews-model/src/main/java/com/heima/model/dto/user/UserRelationDto.java
@@ -10,7 +10,7 @@ public class UserRelationDto {
      * 文章id
      */
     @IdEncrypt
-    private Integer articleId;
+    private Long articleId;
 
     /**
      * 作者id
